Added QOL chat features and changed how messages are displayed
This was a pause in major chat revisions. I added the following QOL
chat features:

 1. Alterting on User Exit
 2. Alterting on User Join
 3. Removed unnecissary urls
 4. Fixed spacing bugs

I also changed how messages are displayed on the front end. It used
to be a text box but I changed it so that messages are appended in
divs into one large div. This will allow a ton of CSS customization
like text user coloring, how messages look, and a ton. This was the
start of my realization that I will have to do the front end :( .