Updated table additions and color'd chat text.
I updated the chat so that each user gets a different color with
the client being forge red and then each additional user being a
different color. This is all done client side. Also updated how
data is dynamically added to the table to actually use the API
and not just manually add the table