Front end form redo Part 1#
This is a inprogress commit of reworking the front end for the
user form. It was done very poorly the first time. I realize
that correct OSS is to branch and do this but I am being a little
lazy and there is no one else actively working on this repo. Now
one to the progress. This replaces any manual <br> and div positioning
with actual css. As I said in the website repo. :' ( I dont lke css