Finalized form update and chat css
I fixed the issue where the css on the form was
not correctly aligned. I also fixed a css issue
where the chat upload and send buttons were not
the right size. Finally I centered the data
in the table columns inline with user feedback.

I realize this took longer than my previous things
and that is because I absolutly stink at css.

:'( I hate css